cipher file cleanup
